server/requirements.txt
# Vidzeo Local - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Text-to-Speech
kokoro>=0.9.4

# Speech Recognition / Transcription
openai-whisper>=20231117

# AI APIs
groq>=0.4.0
openai>=1.0.0

# NSFW Detection
opennsfw2>=0.1.1

# Web Search
duckduckgo-search>=4.0

# Image Processing
Pillow>=10.0.0

# Audio Processing
numpy>=1.24.0
soundfile>=0.12.0
pydub>=0.25.0

# HTTP Requests
requests>=2.31.0

# espeak-ng is required for Kokoro TTS (install separately on Windows)
